requirements.txt
flask==3.0.0
bandit==1.7.5
jinja2==3.1.2
werkzeug==3.0.1
requests==2.31.0
pyyaml==6.0.1

